Add Vehicle connection status to vehicle table view - expansion row
