+++ packages/modules/web_themes/koala/source/src/components/VehicleCard.vue
@@ -16,40 +16,7 @@
           {{ vehicleInfo?.model || 'keine Angabe' }}
         </div>
       </div>
-      <div class="row q-mt-sm">
-        <div class="col">
-          <div class="text-subtitle2">Status:</div>
-          <q-chip
-            v-if="vehicleState.length < 1"
-            label="Nicht zugeordnet"
-            color="primary"
-          >
-          </q-chip>
-          <q-chip
-            v-for="(chargePoint, index) in vehicleState"
-            :key="index"
-            :icon="chargePoint.plugged ? 'power' : 'power_off'"
-            :color="
-              chargePoint.plugged
-                ? chargePoint.charging
-                  ? 'positive'
-                  : 'warning'
-                : 'negative'
-            "
-            :label="chargePoint.name"
-          >
-            <q-tooltip>
-              {{
-                chargePoint.plugged
-                  ? chargePoint.charging
-                    ? 'Lädt'
-                    : 'Angesteckt, lädt nicht'
-                  : 'Nicht angesteckt'
-              }}
-            </q-tooltip>
-          </q-chip>
-        </div>
-      </div>
+      <VehicleConnectionStateIcon :vehicle-id="vehicleId" />
       <div v-if="vehicleSocValue !== null">
         <SliderDouble
           class="q-mt-sm"
@@ -97,6 +64,7 @@ import { useMqttStore } from 'src/stores/mqtt-store';
 import { useQuasar } from 'quasar';
 import SliderDouble from './SliderDouble.vue';
 import ManualSocDialog from './ManualSocDialog.vue';
+import VehicleConnectionStateIcon from './VehicleConnectionStateIcon.vue';
 
 const props = defineProps<{
   vehicleId: number;
@@ -122,9 +90,9 @@ const vehicleSocValue = computed(() => {
   return mqttStore.vehicleSocValue(props.vehicleId);
 });
 
-const vehicleState = computed(() => {
-  return mqttStore.vehicleConnectionState(props.vehicleId);
-});
+// const vehicleState = computed(() => {
+//   return mqttStore.vehicleConnectionState(props.vehicleId);
+// });
 
 const refreshSoc = () => {
   mqttStore.vehicleForceSocUpdate(props.vehicleId);
